refactor(client): change client login page organization
